# swagger.yaml
openapi: 3.0.0
info:
  title: Investment Simulator API
  version: 1.0.0
  description: API para o simulador de investimentos.

servers:
  - url: http://localhost:8000 # Altere a porta se necessário
    description: Servidor de Desenvolvimento

tags:
  - name: User
    description: Operações relacionadas ao usuário

paths:
  /user/signup:
    post:
      tags:
        - User
      summary: Cria um novo usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '200':
          description: Usuário criado com sucesso. Retorna ID e token.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  token:
                    type: string
        '400':
          description: E-mail já em uso ou dados inválidos.

  /user/signin:
    post:
      tags:
        - User
      summary: Autentica um usuário e retorna um token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Login bem-sucedido.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  token:
                    type: string
        '400':
          description: Credenciais inválidas.

  /user/profile:
    get:
      tags:
        - User
      summary: Retorna o perfil do usuário autenticado
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Perfil do usuário.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '403':
          description: Não autorizado (token ausente ou inválido).
    patch:
      tags:
        - User
      summary: Atualiza o perfil do usuário autenticado (nome e/ou senha)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Perfil atualizado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          description: Nenhum dado fornecido para atualização.
        '403':
          description: Não autorizado.

components:
  schemas:
    UserInput:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          example: 'John Doe'
        email:
          type: string
          format: email
          example: 'john.doe@example.com'
        password:
          type: string
          format: password
          example: 'password123'

    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: 'john.doe@example.com'
        password:
          type: string
          format: password
          example: 'password123'

    UserUpdate:
      type: object
      properties:
        name:
          type: string
          description: O novo nome do usuário.
          example: 'Johnathan Doe'
        password:
          type: string
          format: password
          description: A nova senha do usuário.
          example: 'newSecurePassword456'

    UserProfile:
      type: object
      properties:
        _id:
          type: string
          example: '60d0fe4f5311236168a109ca'
        name:
          type: string
          example: 'John Doe'
        email:
          type: string
          format: email
          example: 'john.doe@example.com'

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT